refactor(frontend): improve IssueReviewView sidebar structure and maintainability
- Split DatabaseChangeSection into StagesSection and ChecksSection
- Extract ApprovalStepItem logic into useApprovalStep composable
- Reorder sections to follow popular platform patterns (Stages → Checks → Approval → Labels)
- Remove redundant IssueStatusSection (status is already shown in approval flow)
- Add approval template title display in ApprovalFlowSection header
