+++ frontend/src/components/Plan/components/IssueReviewView/Sidebar/ApprovalFlowSection/ApprovalFlowSection.vue
@@ -2,6 +2,12 @@
   <div>
     <div class="w-full flex flex-row items-center gap-2">
       <h3 class="textlabel">{{ $t("issue.approval-flow.self") }}</h3>
+      <span
+        v-if="approvalTemplate?.title?.trim()"
+        class="text-sm text-control-placeholder"
+      >
+        · {{ approvalTemplate.title }}
+      </span>
       <FeatureBadge :feature="PlanFeature.FEATURE_APPROVAL_WORKFLOW" />
       <RiskLevelIcon :risk-level="issue.riskLevel" />
     </div>
