+++ frontend/src/components/Plan/components/IssueReviewView/Sidebar/StagesSection.vue
@@ -1,0 +1,74 @@
+<template>
+  <div v-if="rollout" class="flex flex-col gap-y-1">
+    <h3 class="textlabel">
+      {{ $t("rollout.stage.self", 2) }}
+    </h3>
+    <div class="flex flex-wrap items-center gap-y-1">
+      <template v-if="rollout.stages.length > 0">
+        <template v-for="(stage, index) in rollout.stages" :key="stage.name">
+          <div
+            class="flex items-center gap-1 cursor-pointer hover:opacity-80 transition-opacity"
+            @click="navigateToStage(stage.name)"
+          >
+            <TaskStatus :status="getStageStatus(stage)" size="small" disabled />
+            <EnvironmentV1Name
+              :environment="getEnvironmentEntity(stage.environment)"
+              :link="false"
+              :null-environment-placeholder="'Null'"
+            />
+          </div>
+          <span
+            v-if="index < rollout.stages.length - 1"
+            class="mx-2 text-control-placeholder"
+          >
+            →
+          </span>
+        </template>
+      </template>
+      <span v-else class="text-sm text-control-placeholder">
+        {{ $t("common.no-data") }}
+      </span>
+    </div>
+  </div>
+</template>
+
+<script setup lang="ts">
+import { useRouter } from "vue-router";
+import TaskStatus from "@/components/Rollout/kits/TaskStatus.vue";
+import { EnvironmentV1Name } from "@/components/v2";
+import { PROJECT_V1_ROUTE_PLAN_ROLLOUT_STAGE } from "@/router/dashboard/projectV1";
+import { useCurrentProjectV1, useEnvironmentV1Store } from "@/store";
+import {
+  extractPlanUIDFromRolloutName,
+  extractProjectResourceName,
+  getStageStatus,
+} from "@/utils";
+import { usePlanContext } from "../../../logic";
+
+const { rollout } = usePlanContext();
+const environmentStore = useEnvironmentV1Store();
+const router = useRouter();
+const { project } = useCurrentProjectV1();
+
+const getEnvironmentEntity = (environmentName: string) => {
+  return environmentStore.getEnvironmentByName(environmentName);
+};
+
+const navigateToStage = (stageName: string) => {
+  if (!rollout?.value) return;
+
+  const planId = extractPlanUIDFromRolloutName(rollout.value.name);
+  const stageId = stageName.split("/").pop();
+
+  if (!planId || !stageId) return;
+
+  router.push({
+    name: PROJECT_V1_ROUTE_PLAN_ROLLOUT_STAGE,
+    params: {
+      projectId: extractProjectResourceName(project.value.name),
+      planId,
+      stageId,
+    },
+  });
+};
+</script>
