+++ frontend/src/locales/zh-CN.json
@@ -1148,9 +1148,6 @@
     },
     "review": {
       "self": "审批",
-      "approved": "已批准",
-      "rejected": "已驳回",
-      "under-review": "等待审批",
       "comment-description": "提交反馈，不进行批准操作",
       "approve-description": "提交反馈并批准此工单",
       "reject-description": "提交反馈并要求修改"
